rename auth users utils.js
