+++ client/src/pages/Login.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useGoogleLogin } from '@react-oauth/google'
 import { useNavigate } from 'react-router-dom'
 import axios from 'axios'
-import { fetchUserDetails, createUserIfNotFound } from '../auth/utils'
+import { fetchUserDetails, createUserIfNotFound } from '../auth-users/utils'
 import Hero from '../components/Hero.jsx'
 import { FormattedMessage } from 'react-intl' // To show localized strings
